+++ src/Demos/MusicWinFromsDemo/Program.cs
@@ -9,10 +9,6 @@ namespace MusicWinFromsDemo
         {
             ApplicationConfiguration.Initialize();
 
-#if LVGLSHARP_FORMS
-            DemoRuntimeConfiguration.Configure();
-#endif
-
             Application.Run(new frmMusicDemo());
         }
     }